// src/generated/AppLifecyclePhase.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AppLifecyclePhase = "booting" | "shell_ready" | "database_initializing" | "core_ready" | "proxy_starting" | "running" | "recovery_required" | "database_error" | "port_conflict" | "proxy_error" | "shutting_down" | "exited";